requirements.txt
tweepy==3.10.0
numpy==1.19.4
pandas==1.1.3
Pillow==8.1.0
seaborn 
matplotlib
streamlit
google-cloud-bigquery